refactor: Changed the project structure to hexagonal. Parallel study to verify optimization.
