+++ src/main/java/br/com/lasbr/adopetapi/adapter/controller/PetController.java
@@ -1,7 +1,7 @@
-package br.com.lasbr.adopetapi.controller;
+package br.com.lasbr.adopetapi.adapter.controller;
 
-import br.com.lasbr.adopetapi.entity.Pet;
-import br.com.lasbr.adopetapi.service.PertService;
+import br.com.lasbr.adopetapi.adapter.data.entities.Pet;
+import br.com.lasbr.adopetapi.domain.service.PertService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
