+++ src/main/java/br/com/lasbr/adopetapi/adapter/data/entities/Shelter.java
@@ -1,14 +1,10 @@
-package br.com.lasbr.adopetapi.entity;
+package br.com.lasbr.adopetapi.adapter.data.entities;
 
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
 
 import java.util.ArrayList;
 import java.util.List;
